refactor: centralize extraction helpers
